+++ docs/CHANGELOG.md
@@ -12,13 +12,13 @@ nav_order: 5
 
 ## 4.0.0
 
-* BREAKING: Require [non-EOL](https://endoflife.date/rails) Rails (`>= 7.0.0`).
+* BREAKING: Require [non-EOL](https://endoflife.date/rails) Rails (`>= 7.1.0`).
 
     *Joel Hawksley*
 
-* BREAKING: Require [non-EOL](https://www.ruby-lang.org/en/downloads/branches/) Ruby (`>= 3.1.0`).
+* BREAKING: Require [non-EOL](https://www.ruby-lang.org/en/downloads/branches/) Ruby (`>= 3.2.0`).
 
-    * Joel Hawksley*
+    *Joel Hawksley*
 
 * Add basic internal testing for memory allocations.