+++ docs/CHANGELOG.md
@@ -10,6 +10,8 @@ nav_order: 5
 
 ## main
 
+## 3.15.0
+
 * Add basic internal testing for memory allocations.
 
     *Joel Hawksley*